\name{ch7tps}
\docType{data}
\alias{ch7tps}
\title{TPS Data for the Vectorized Root System of a Young Oak Tree}
\description{This dataset obtained with Data Analysis of Root Tracings gives the correspondence between the observation numbers, dates and colour codes used to perform a root system architecture analysis of a young oak tree (\emph{Quercus robur} L.) growing in a vertical rhizobox using DART.  
}
\usage{ch7tps}
\format{A data frame containing the following columns:
\itemize{
\item \bold{Num}: the identification number of each observation date.
\item \bold{Date}: the root system age at each observation date (the unit of time is defined by the experimenter). In ch7tps, data are expressed in days.
\item \bold{CoulR}: the red component of the colour chosen for the vectorization of the root system at each observation date.
\item \bold{CoulG}: the green component of the colour chosen for the vectorization of the root system at each observation date.
\item \bold{CoulB}: the blue component of the colour chosen for the vectorization of the root system at each observation date.}
}
\source{\url{www6.paca.inra.fr/psh/Modeles-et-logiciels/Dart}
}
\references{
Pages L., Serra V. (1994) Growth and branching of the taproot of young oak trees - a dynamic study, \emph{Journal of Experimental Botany}, \bold{45}, 1327--1334.

Le Bot J., Serra V., Fabre J., Draye X., Adamowicz S., Pages L. (2010) DART: a software to analyse root system architecture and development from captured images, \emph{Plant and Soil}, \bold{326}, 261--273.
}
\keyword{datasets}